+++ test/lib/rules/operation-payload-put.js
@@ -29,7 +29,8 @@ describe('operation-payload-put', () => {
               }
             }
           ]
-        }
+        },
+        parameters: []
       }
     }
   };